refactor(reposicao): split CicloHojePanel (728→159) em hook + componentes
Extrai a lógica do painel para o hook useCicloHoje (seleção, aprovação em
lote/automática, memos de elegíveis/grupos/revisão manual, invalidate) e
quebra o JSX em componentes controlados (PedidoRow + PedidoRowCells,
FiltersToolbar, AutoApproveDialog, BatchActionsBar). inferConfianca isolada
em confianca.ts (pura). ALL re-exportado do arquivo principal (preserva
import de AdminReposicaoSessaoPedidos). Comportamento preservado 1:1; +25 testes.
